Add mistyped_uneval example
A minimal example that generates mistyped unevaluation warnings.
This is caused by term_gen incorrectly using recursive bindings as the argument to a function.
A simple fix is to simply disallow the use of recursive bindings outside of application heads.
A better fix would be to add some form of termination checking.